refactor(site/src): move LoadMoreSentinel to shared components
