+++ site/src/pages/AgentsPage/components/ChatsSidebar/chats/ChatsPanel.tsx
@@ -28,6 +28,7 @@ import { Button } from "#/components/Button/Button";
 import { FeatureStageBadge } from "#/components/FeatureStageBadge/FeatureStageBadge";
 import { ProductLogo } from "#/components/Icons/ProductLogo";
 import { Kbd, KbdGroup } from "#/components/Kbd/Kbd";
+import { LoadMoreSentinel } from "#/components/LoadMoreSentinel/LoadMoreSentinel";
 import { ScrollArea } from "#/components/ScrollArea/ScrollArea";
 import { Skeleton } from "#/components/Skeleton/Skeleton";
 import { cn } from "#/utils/cn";
@@ -58,7 +59,6 @@ import {
 	getSectionToggleTestId,
 	PINNED_SECTION_KEY,
 } from "./ChatSectionHeader";
-import { LoadMoreSentinel } from "./LoadMoreSentinel";
 import { UserSidebarFooter } from "./UserSidebarFooter";
 
 const UNREAD_SECTION_KEY = "Unread";
